// web/src/api/generated/models/catalogManagementModifierGroupResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * POS Cafe API
 * Backend API for Cafe Point of Sale System built with Vertical Slice Architecture and Idiomatic Go.
 * OpenAPI spec version: 1.0
 */
import type { CatalogManagementModifierOptionResponse } from './catalogManagementModifierOptionResponse';

export interface CatalogManagementModifierGroupResponse {
  default_option_ids?: string[];
  id?: string;
  max_selections?: number;
  min_selections?: number;
  name?: string;
  options?: CatalogManagementModifierOptionResponse[];
  retired?: boolean;
  retired_at?: string;
  retirement_note?: string;
  retirement_reason?: string;
}